requirements.txt
fastapi
uvicorn
motor
python-multipart
pillow
pytesseract
